revert SelectExpiringLiquidityAmount unneeded type change

+++ internal/infrastructure/db/postgres/sqlc/query.sql
@@ -263,7 +263,7 @@ WHERE swept = false
   AND spent = false
   AND unrolled = false
   AND expires_at > @after
-  AND (@before::bigint <= 0 OR expires_at < @before);
+  AND (@before <= 0 OR expires_at < @before);
 
 -- name: SelectRecoverableLiquidityAmount :one
 SELECT COALESCE(SUM(amount), 0)::bigint AS amount
